app/src/main/java/com/example/todoapp/data/NoteDatabase.kt
package com.example.todoapp.data

import android.content.Context
import androidx.room.Database
import androidx.room.Room
import androidx.room.RoomDatabase
import com.example.todoapp.data.models.Note
import com.example.todoapp.utils.AppConstraints

@Database(entities = [Note::class], version = AppConstraints.DATABASE_VERSION, exportSchema = false)
abstract class NoteDatabase: RoomDatabase() {
    abstract fun createNoteDAO(): NoteDAO

    companion object {
        // Singleton prevents multiple instances of database opening at the
        // same time.
        @Volatile
        private var INSTANCE: NoteDatabase? = null

        fun getDatabase(context: Context): NoteDatabase {
            // if the INSTANCE is not null, then return it,
            // if it is, then create the database
            return INSTANCE ?: synchronized(this) {
                val instance = Room.databaseBuilder(
                    context.applicationContext,
                    NoteDatabase::class.java,
                    AppConstraints.DATABASE_NAME
                ).build()
                INSTANCE = instance
                // return instance
                instance
            }
        }
    }
}